# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Vuokraappi Backend API
  description: |
    REST API for Vuokraappi – a Finnish rental apartment platform serving both
    tenants and landlords. User identities are managed via Microsoft Entra ID;
    the backend stores a local user record and a role-specific profile
    (tenant or landlord) for each registered user.
  version: 0.0.1-SNAPSHOT
  contact:
    name: Vuokraappi
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.dev.vuokraappi.fi
    description: Development environment (Azure)
...
paths:
...
  /v1/users/{userId}:
    get:
      tags:
        - Users
      summary: Get user by ID
      operationId: getUser
      parameters:
        - name: userId
          in: path
          required: true
          description: UUID of the user
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags:
        - Users
      summary: Update user profile
      description: |
        Partially updates user information. All fields are optional.
        Supports both tenant-specific fields (occupation, monthlyIncome, etc.)
        and landlord-specific fields (companyName, businessId, bankAccount).
      operationId: updateUser
      parameters:
        - name: userId
          in: path
          required: true
          description: UUID of the user to update
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
            example:
              firstName: "Matti"
              lastName: "Virtanen"
              phoneNumber: "+358401234567"
              email: "matti.virtanen@example.com"
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Users
      summary: Delete a user
      description: |
        Permanently deletes a user from both the local database and Microsoft Entra ID.
        Cascade deletes the associated Tenant or Landlord profile automatically.
        If the user is not found in Entra ID during deletion, the operation still succeeds
        (idempotent Entra ID removal).
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          required: true
          description: UUID of the user to delete
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '204':
          description: User deleted successfully (no content)
        '404':
          description: User not found in the database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: 404
                message: "User with ID 550e8400-e29b-41d4-a716-446655440000 not found"
                timestamp: "2024-02-16T10:30:00"
        '502':
          description: Microsoft Entra ID (Graph API) communication failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/apartments/{id}:
    get:
      tags:
        - Apartments
      summary: Get apartment by ID
      operationId: getApartmentById
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the apartment
          schema:
            type: string
            format: uuid
          example: "770e8400-e29b-41d4-a716-446655440002"
      responses:
        '200':
          description: Apartment found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApartmentResponse'
        '404':
          description: Apartment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: 404
                message: "Apartment not found with id: 770e8400-e29b-41d4-a716-446655440002"
                timestamp: "2024-02-16T10:30:00"
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags:
        - Apartments
      summary: Update an apartment
      description: |
        Partially updates an apartment. Only the fields present in the request body are updated.
        The authenticated user must be the owner of the apartment.
      operationId: updateApartment
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the apartment to update
          schema:
            type: string
            format: uuid
          example: "770e8400-e29b-41d4-a716-446655440002"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApartmentRequest'
            example:
              rent: 1100.00
      responses:
        '200':
          description: Apartment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApartmentResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Authenticated user is not the owner of the apartment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: 403
                message: "You can only update your own apartments"
                timestamp: "2024-02-16T10:30:00"
        '404':
          description: Apartment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Apartments
      summary: Delete an apartment
      description: |
        Deletes an apartment. The authenticated user must be the owner of the apartment.
      operationId: deleteApartment
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the apartment to delete
          schema:
            type: string
            format: uuid
          example: "770e8400-e29b-41d4-a716-446655440002"
      responses:
        '204':
          description: Apartment deleted successfully (no content)
        '403':
          description: Authenticated user is not the owner of the apartment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: 403
                message: "You can only delete your own apartments"
                timestamp: "2024-02-16T10:30:00"
        '404':
          description: Apartment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ---------------------------------------------------------------------------
  # Apartment Images
  # ---------------------------------------------------------------------------
...
  /v1/apartments/{apartmentId}/swipes:
    post:
      tags:
        - Swipes & Matches
      summary: Tenant swipes on an apartment
      description: |
        Records the authenticated tenant's swipe decision on an apartment.
        Re-swiping the same apartment updates the previous decision.
        If `liked=true` and the apartment's owner has already liked this tenant
        for this apartment, a match is created automatically and returned.
      operationId: swipeApartment
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - name: apartmentId
          in: path
          required: true
          description: UUID of the apartment
          schema:
            type: string
            format: uuid
          example: "770e8400-e29b-41d4-a716-446655440002"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SwipeRequest'
            example:
              liked: true
      responses:
        '200':
          description: Swipe recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantSwipeResponse'
        '400':
          description: Validation error – missing `liked` field
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not authenticated, or authenticated user is not a tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Apartment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Swipes & Matches
      summary: Remove a tenant's swipe on an apartment
      description: |
        Removes the authenticated tenant's own swipe decision on the apartment.
        Does not affect a match that may have already been created – use
        `DELETE /v1/matches/{matchId}` to unmatch.
      operationId: removeApartmentSwipe
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - name: apartmentId
          in: path
          required: true
          description: UUID of the apartment
          schema:
            type: string
            format: uuid
          example: "770e8400-e29b-41d4-a716-446655440002"
      responses:
        '204':
          description: Swipe removed (no content)
        '403':
          description: Not authenticated, or authenticated user is not a tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/apartments/{apartmentId}/matches:
    get:
      tags:
        - Swipes & Matches
      summary: List matches for an apartment (landlord use)
      description: |
        Returns all matches for the given apartment. The authenticated user
        must be the owner of the apartment.
      operationId: listApartmentMatches
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - name: apartmentId
          in: path
          required: true
          description: UUID of the apartment
          schema:
            type: string
            format: uuid
          example: "770e8400-e29b-41d4-a716-446655440002"
      responses:
        '200':
          description: List of matches (empty if none)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchResponse'
        '403':
          description: Not authenticated, or authenticated user is not the owner of the apartment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Apartment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/apartments/{apartmentId}/tenants/{tenantId}/swipes:
    post:
      tags:
        - Swipes & Matches
      summary: Landlord swipes on a tenant
      description: |
        Records the authenticated landlord's swipe decision on a tenant, scoped
        to a specific apartment the landlord owns. Re-swiping the same tenant
        for the same apartment updates the previous decision.
        If `liked=true` and the tenant has already liked this apartment, a
        match is created automatically and returned.
      operationId: swipeTenant
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - name: apartmentId
          in: path
          required: true
          description: UUID of the apartment (must be owned by the authenticated landlord)
          schema:
            type: string
            format: uuid
          example: "770e8400-e29b-41d4-a716-446655440002"
        - name: tenantId
          in: path
          required: true
          description: UUID of the tenant being swiped on
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SwipeRequest'
            example:
              liked: true
      responses:
        '200':
          description: Swipe recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LandlordSwipeResponse'
        '400':
          description: Validation error – missing `liked` field
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not authenticated, or authenticated user does not own the apartment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Apartment or tenant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Swipes & Matches
      summary: Remove a landlord's swipe on a tenant
      description: |
        Removes the authenticated landlord's own swipe decision on the tenant
        for the given apartment. Does not affect a match that may have already
        been created – use `DELETE /v1/matches/{matchId}` to unmatch.
      operationId: removeTenantSwipe
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - name: apartmentId
          in: path
          required: true
          description: UUID of the apartment (must be owned by the authenticated landlord)
          schema:
            type: string
            format: uuid
          example: "770e8400-e29b-41d4-a716-446655440002"
        - name: tenantId
          in: path
          required: true
          description: UUID of the tenant
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '204':
          description: Swipe removed (no content)
        '403':
          description: Not authenticated, or authenticated user does not own the apartment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Apartment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/apartments/{apartmentId}/tenant-swipes:
    get:
      tags:
        - Swipes & Matches
      summary: List a landlord's own swipes for an apartment
      description: |
        Returns all of the authenticated landlord's swipe decisions on tenants
        for the given apartment (i.e. tenants already handled). The
        authenticated user must be the owner of the apartment.
      operationId: listTenantSwipes
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - name: apartmentId
          in: path
          required: true
          description: UUID of the apartment
          schema:
            type: string
            format: uuid
          example: "770e8400-e29b-41d4-a716-446655440002"
      responses:
        '200':
          description: List of swipes (empty if none)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LandlordSwipeResponse'
        '403':
          description: Not authenticated, or authenticated user is not the owner of the apartment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Apartment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/tenants/{tenantId}/swipes:
    get:
      tags:
        - Swipes & Matches
      summary: List a tenant's own swipes
      description: |
        Returns all of the authenticated tenant's swipe decisions on apartments
        (i.e. apartments already handled). The authenticated user must be the
        tenant themself.
      operationId: listSwipes
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          description: UUID of the tenant (same as the user's UUID)
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: List of swipes (empty if none)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TenantSwipeResponse'
        '403':
          description: Not authenticated, or authenticated user is not the tenant themself
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    UserUpdateRequest:
      type: object
      description: All fields are optional. Only supplied fields are updated.
      properties:
        firstName:
          type: string
          maxLength: 100
          example: "Matti"
        lastName:
          type: string
          maxLength: 100
          example: "Virtanen"
        phoneNumber:
          type: string
          maxLength: 20
          example: "+358401234567"
        email:
          type: string
          format: email
          maxLength: 255
          example: "matti.virtanen@example.com"
        introduction:
          type: string
          maxLength: 2000
          example: "Rauhallinen ja siisti vuokralainen."
        dateOfBirth:
          type: string
          format: date
          example: "1990-05-15"
        personalIdentityCode:
          type: string
          maxLength: 25
          example: "150590-1234"
        # Tenant-specific
        occupation:
          type: string
          maxLength: 100
          example: "Opiskelija"
        monthlyIncome:
          type: integer
          example: 2500
        currentAddress:
          type: string
          maxLength: 255
          example: "Mannerheimintie 1, Helsinki"
        pet:
          type: string
          maxLength: 100
          example: "Kissa"
        # Landlord-specific
        companyName:
          type: string
          maxLength: 200
          example: "Kiinteistö Oy Esimerkki"
        businessId:
          type: string
          maxLength: 50
          example: "1234567-8"
        bankAccount:
          type: string
          maxLength: 100
          example: "FI21 1234 5600 0007 85"
    ApartmentRequest:
      type: object
      description: Request body for creating or partially updating an apartment. All fields are optional for PATCH.
      properties:
        zipcode:
          type: string
          description: Finnish postal code
          example: "00530"
        streetAddress:
          type: string
          example: "Fleminginkatu 12"
        city:
          type: string
          example: "Helsinki"
        region:
          type: string
          example: "Uusimaa"
        size:
          type: number
          format: double
          description: Apartment size in square metres
          example: 42.5
        longitude:
          type: number
          format: double
          example: 24.9534
        latitude:
          type: number
          format: double
          example: 60.1828
        rent:
          type: number
          format: decimal
          description: Monthly rent in euros
          example: 1050.00
...
    SwipeRequest:
      type: object
      required:
        - liked
      properties:
        liked:
          type: boolean
          description: |
            `true` = interested (like), `false` = not interested (pass).
            Re-swiping the same target updates the previous decision.
          example: true
...
    UserDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        role:
          $ref: '#/components/schemas/UserRole'
        firstName:
          type: string
          nullable: true
          example: "Matti"
        lastName:
          type: string
          nullable: true
          example: "Virtanen"
        phoneNumber:
          type: string
          nullable: true
          example: "+358401234567"
        email:
          type: string
          nullable: true
          example: "matti.virtanen@example.com"
        introduction:
          type: string
          nullable: true
          example: "Rauhallinen ja siisti vuokralainen."
        dateOfBirth:
          type: string
          format: date
          nullable: true
          example: "1990-05-15"
        personalIdentityCode:
          type: string
          nullable: true
          example: "150590-1234"
        createdAt:
          type: string
          format: date-time
          example: "2024-02-16T10:30:00"
        updatedAt:
          type: string
          format: date-time
          example: "2024-03-01T08:00:00"
        isActive:
          type: boolean
          example: true
        # Tenant-specific (null for landlords)
        occupation:
          type: string
          nullable: true
          example: "Opiskelija"
        monthlyIncome:
          type: integer
          nullable: true
          example: 2500
        currentAddress:
          type: string
          nullable: true
          example: "Mannerheimintie 1, Helsinki"
        pet:
          type: string
          nullable: true
          example: "Kissa"
        # Landlord-specific (null for tenants)
        companyName:
          type: string
          nullable: true
          example: "Kiinteistö Oy Esimerkki"
        businessId:
          type: string
          nullable: true
          example: "1234567-8"
        bankAccount:
          type: string
          nullable: true
          example: "FI21 1234 5600 0007 85"
...
    ApartmentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "770e8400-e29b-41d4-a716-446655440002"
        ownerId:
          type: string
          format: uuid
          description: UUID of the landlord who owns the apartment
          example: "660e8400-e29b-41d4-a716-446655440001"
        ownerName:
          type: string
          description: Full name of the landlord (firstName + lastName)
          example: "Matti Virtanen"
        zipcode:
          type: string
          example: "00530"
        streetAddress:
          type: string
          example: "Fleminginkatu 12"
        city:
          type: string
          example: "Helsinki"
        region:
          type: string
          example: "Uusimaa"
        size:
          type: number
          format: double
          example: 42.5
        longitude:
          type: number
          format: double
          example: 24.9534
        latitude:
          type: number
          format: double
          example: 60.1828
        rent:
          type: number
          format: decimal
          example: 1050.00
        ownerIdentityVerified:
          type: boolean
          description: Whether the landlord has completed strong identity verification (Signicat)
          example: true
...
    TenantSwipeResponse:
      type: object
      description: |
        Response for a tenant's swipe action. `matched`/`matchId` indicate whether
        this swipe resulted in a mutual match – these are not populated in
        listing calls, only in the response of the swipe action itself.
      properties:
        id:
          type: string
          format: uuid
          example: "aa0e8400-e29b-41d4-a716-446655440005"
        apartmentId:
          type: string
          format: uuid
          example: "770e8400-e29b-41d4-a716-446655440002"
        liked:
          type: boolean
          example: true
        matched:
          type: boolean
          nullable: true
          example: true
        matchId:
          type: string
          format: uuid
          nullable: true
          example: "990e8400-e29b-41d4-a716-446655440004"
        createdAt:
          type: string
          format: date-time
          example: "2024-02-16T10:30:00"
...
    MatchResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "990e8400-e29b-41d4-a716-446655440004"
        tenantId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        tenantName:
          type: string
          example: "Aino Mäkinen"
        apartmentId:
          type: string
          format: uuid
          example: "770e8400-e29b-41d4-a716-446655440002"
        apartmentAddress:
          type: string
          example: "Fleminginkatu 12, Helsinki"
        landlordId:
          type: string
          format: uuid
          example: "660e8400-e29b-41d4-a716-446655440001"
        matchedAt:
          type: string
          format: date-time
          example: "2024-02-16T10:30:00"
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          description: HTTP status code
          example: 400
        message:
          type: string
          description: Human-readable error message
          example: "Validation failed"
        timestamp:
          type: string
          format: date-time
          description: Time when the error occurred
          example: "2024-02-16T10:30:00"
        errors:
          type: object
          description: >
            Field-level validation errors (only present for 400 responses).
            Keys are field names, values are error messages.
          additionalProperties:
            type: string
          example:
            role: "Role must be either 'tenant' or 'landlord'"

    # -------------------------------------------------------------------------
    # Enums
    # -------------------------------------------------------------------------
    UserRole:
      type: string
      description: The user's role in the platform
      enum:
        - TENANT
        - LANDLORD
      example: "TENANT"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token validated by the API Gateway (dev/prod environments).
        The gateway extracts the user's Entra ID Object ID from the `oid` claim
        and forwards it to the backend service.
    BasicAuth:
      type: http
      scheme: basic
      description: |
        Used in the local development environment only (email-based login).
security:
  - BearerAuth: []